refactor(core): decouple Lightpath, Flexgrind and Allocation
