+++ application/core/flexgrid.h
@@ -3,8 +3,6 @@
 #include <cstdint>
 #include <memory>
 
-#include "lightpath.h"
-
 namespace core {
 class Flexgrid final {
  public:
@@ -26,11 +24,7 @@ class Flexgrid final {
 
   Flexgrid& operator=(Flexgrid&&) noexcept;
 
-  void Allocate(const Lightpath&);
-
   void Allocate(const Unit);
-
-  void Deallocate(const Lightpath&);
 
   void Deallocate(const Unit);
